+++ src/Components/Cartbutton/CartButton.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './styles.scss';
+
+const CartButton = () => {
+  return (
+    <div className="cart">
+      <img src="shopping-cart.svg" alt="Cart" />
+    </div>
+  );
+};
+
+export default CartButton;
