+++ src/Components/Cartbutton/index.js
@@ -1,12 +1,1 @@
-import React from 'react';
-import './styles.scss';
-
-const CartButton = () => {
-  return (
-    <div className="cart">
-      <img src="shopping-cart.svg" alt="Cart" />
-    </div>
-  );
-};
-
-export default CartButton;
+export { default } from './CartButton';
