+++ src/Components/Footer/Footer.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import './styles.scss';
+
+const Footer = () => {
+  return (
+    <footer className="footer">
+      <div>
+        <img src="AcademyLogo.jpg" alt="" className="academyLogo" />
+        <p>
+          <i>Ecommerce created during Wizeline’s Academy React Bootcamp</i>
+        </p>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;
