+++ src/Components/Footer/index.js
@@ -1,17 +1,1 @@
-import React from 'react';
-import './styles.scss';
-
-const Footer = () => {
-  return (
-    <footer className="footer">
-      <div>
-        <img src="AcademyLogo.jpg" alt="" className="academyLogo" />
-        <p>
-          <i>Ecommerce created during Wizeline’s Academy React Bootcamp</i>
-        </p>
-      </div>
-    </footer>
-  );
-};
-
-export default Footer;
+export { default } from './Footer';
