+++ src/Components/SearchBox/SearchBox.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './styles.scss';
+
+const SearchBox = () => {
+  return (
+    <div className="SearchBox">
+      <input type="search" placeholder="search for something" />
+    </div>
+  );
+};
+
+export default SearchBox;
