+++ src/Components/SearchBox/index.js
@@ -1,12 +1,1 @@
-import React from 'react';
-import './styles.scss';
-
-const SearchBox = () => {
-  return (
-    <div className="SearchBox">
-      <input type="search" placeholder="search for something" />
-    </div>
-  );
-};
-
-export default SearchBox;
+export { default } from './SearchBox';
